refactor perception_reproducer
