+++ planning/planning_debug_tools/scripts/perception_replayer/perception_reproducer.py
@@ -17,144 +17,31 @@
 import argparse
 import copy
 import math
-import os
-from subprocess import CalledProcessError
-from subprocess import check_output
-import time
 
-from autoware_auto_perception_msgs.msg import DetectedObjects
-from autoware_auto_perception_msgs.msg import PredictedObjects
-from autoware_auto_perception_msgs.msg import TrackedObjects
-from autoware_auto_perception_msgs.msg import TrafficSignalArray
-from geometry_msgs.msg import Quaternion
 from nav_msgs.msg import Odometry
 import numpy as np
-import psutil
+from perception_replayer_common import PerceptionReplayerCommon
 import rclpy
-from rclpy.node import Node
-from rclpy.serialization import deserialize_message
-import rosbag2_py
-from rosbag2_py import StorageFilter
-from rosidl_runtime_py.utilities import get_message
-from sensor_msgs.msg import PointCloud2
-from sensor_msgs.msg import PointField
-from tf_transformations import euler_from_quaternion
-from tf_transformations import quaternion_from_euler
+from utils import calc_squared_distance
+from utils import create_empty_pointcloud
+from utils import get_quaternion_from_yaw
+from utils import get_yaw_from_quaternion
 
 
-def get_rosbag_options(path, serialization_format="cdr"):
-    storage_options = rosbag2_py.StorageOptions(uri=path, storage_id="sqlite3")
-
-    converter_options = rosbag2_py.ConverterOptions(
-        input_serialization_format=serialization_format,
-        output_serialization_format=serialization_format,
-    )
-
-    return storage_options, converter_options
-
-
-def open_reader(path: str):
-    storage_options, converter_options = get_rosbag_options(path)
-    reader = rosbag2_py.SequentialReader()
-    reader.open(storage_options, converter_options)
-    return reader
-
-
-def calc_squared_distance(p1, p2):
-    return math.sqrt((p1.x - p2.x) ** 2 + (p1.y - p2.y) ** 2)
-
-
-def create_empty_pointcloud(timestamp):
-    pointcloud_msg = PointCloud2()
-    pointcloud_msg.header.stamp = timestamp
-    pointcloud_msg.header.frame_id = "map"
-    pointcloud_msg.height = 1
-    pointcloud_msg.is_dense = True
-    pointcloud_msg.point_step = 16
-    field_name_vec = ["x", "y", "z"]
-    offset_vec = [0, 4, 8]
-    for field_name, offset in zip(field_name_vec, offset_vec):
-        field = PointField()
-        field.name = field_name
-        field.offset = offset
-        field.datatype = 7
-        field.count = 1
-        pointcloud_msg.fields.append(field)
-    return pointcloud_msg
-
-
-def get_yaw_from_quaternion(orientation):
-    orientation_list = [
-        orientation.x,
-        orientation.y,
-        orientation.z,
-        orientation.w,
-    ]
-    return euler_from_quaternion(orientation_list)[2]
-
-
-def get_quaternion_from_yaw(yaw):
-    q = quaternion_from_euler(0, 0, yaw)
-    orientation = Quaternion()
-    orientation.x = q[0]
-    orientation.y = q[1]
-    orientation.z = q[2]
-    orientation.w = q[3]
-    return orientation
-
-
-class PerceptionReproducer(Node):
+class PerceptionReproducer(PerceptionReplayerCommon):
     def __init__(self, args):
-        super().__init__("perception_reproducer")
-        self.args = args
+        super().__init__(args, "perception_reproducer")
 
-        # subscriber
+        # additional subscriber
         self.sub_odom = self.create_subscription(
             Odometry, "/localization/kinematic_state", self.on_odom, 1
         )
 
-        # publisher
-        if self.args.detected_object:
-            self.objects_pub = self.create_publisher(
-                DetectedObjects, "/perception/object_recognition/detection/objects", 1
-            )
-        elif self.args.tracked_object:
-            self.objects_pub = self.create_publisher(
-                TrackedObjects, "/perception/object_recognition/tracking/objects", 1
-            )
-        else:
-            self.objects_pub = self.create_publisher(
-                PredictedObjects, "/perception/object_recognition/objects", 1
-            )
-
-        self.pointcloud_pub = self.create_publisher(
-            PointCloud2, "/perception/obstacle_segmentation/pointcloud", 1
-        )
-        self.traffic_signals_pub = self.create_publisher(
-            TrafficSignalArray, "/perception/traffic_light_recognition/traffic_signals", 1
-        )
-
         self.ego_pose_idx = None
         self.ego_pose = None
-
         self.prev_traffic_signals_msg = None
 
-        self.rosbag_objects_data = []
-        self.rosbag_ego_odom_data = []
-        self.rosbag_traffic_signals_data = []
-
-        # load rosbag
-        print("Stared loading rosbag")
-        if os.path.isdir(args.bag):
-            for bag_file in sorted(os.listdir(args.bag)):
-                self.load_rosbag(args.bag + "/" + bag_file)
-        else:
-            self.load_rosbag(args.bag)
-        print("Ended loading rosbag")
-
-        # wait for ready to publish/subscribe
-        time.sleep(1.0)
-
+        # start timer callback
         self.timer = self.create_timer(0.01, self.on_timer)
         print("Start timer callback")
 
@@ -162,23 +49,9 @@ class PerceptionReproducer(Node):
         self.ego_pose = odom.pose.pose
 
     def on_timer(self):
-        # kill node if required
-        kill_process_name = None
-        if self.args.detected_object:
-            kill_process_name = "dummy_perception_publisher_node"
-        elif self.args.tracked_object:
-            kill_process_name = "multi_object_tracker"
-        else:
-            kill_process_name = "map_based_prediction"
-        if kill_process_name:
-            try:
-                pid = check_output(["pidof", kill_process_name])
-                process = psutil.Process(int(pid[:-1]))
-                process.terminate()
-            except CalledProcessError:
-                pass
-
         timestamp = self.get_clock().now().to_msg()
+
+        self.kill_online_perception_node()
 
         if self.args.detected_object:
             pointcloud_msg = create_empty_pointcloud(timestamp)
@@ -280,36 +153,6 @@ class PerceptionReproducer(Node):
                 break
 
         return objects_data, traffic_signals_data
-
-    def load_rosbag(self, rosbag2_path: str):
-        reader = open_reader(str(rosbag2_path))
-
-        topic_types = reader.get_all_topics_and_types()
-        # Create a map for quicker lookup
-        type_map = {topic_types[i].name: topic_types[i].type for i in range(len(topic_types))}
-
-        objects_topic = (
-            "/perception/object_recognition/detection/objects"
-            if self.args.detected_object
-            else "/perception/object_recognition/tracking/objects"
-            if self.args.tracked_object
-            else "/perception/object_recognition/objects"
-        )
-        ego_odom_topic = "/localization/kinematic_state"
-        traffic_signals_topic = "/perception/traffic_light_recognition/traffic_signals"
-        topic_filter = StorageFilter(topics=[objects_topic, ego_odom_topic, traffic_signals_topic])
-        reader.set_filter(topic_filter)
-
-        while reader.has_next():
-            (topic, data, stamp) = reader.read_next()
-            msg_type = get_message(type_map[topic])
-            msg = deserialize_message(data, msg_type)
-            if topic == objects_topic:
-                self.rosbag_objects_data.append((stamp, msg))
-            if topic == ego_odom_topic:
-                self.rosbag_ego_odom_data.append((stamp, msg))
-            if topic == traffic_signals_topic:
-                self.rosbag_traffic_signals_data.append((stamp, msg))
 
 
 if __name__ == "__main__":
